add logout link, update login page style, refactor the project structure
