--- src/main/resources/webapp/index.html
@@ -1,112 +1,0 @@
-<!--
-- Description: 	The responsive webapp for videos
-- Date:			21-Aug-2018
-- Author: 		Domi Yang (domi_yang@hotmail.com)
----------------------------------------------------------------------------
-   Copyright 2018 Domi Yang (domi_yang@hotmail.com)
-
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
-
-       http://www.apache.org/licenses/LICENSE-2.0
-
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
----------------------------------------------------------------------------
--->
-
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <!-- The above 3 meta tags *must* come first in the head; any other head content must come *after* these tags -->
-	
-	<!-- no caching at client -->
-	<meta http-equiv="Cache-Control" content="no-cache, no-store, must-revalidate" />
-	<meta http-equiv="Pragma" content="no-cache" />
-	<meta http-equiv="Expires" content="0" />	
-
-    <title>the mp4 site</title>
-    <!-- Bootstrap -->
-    <link href="css/bootstrap.min.css" rel="stylesheet" />
-    <!-- HTML5 shim and Respond.js for IE8 support of HTML5 elements and media queries -->
-    <!-- WARNING: Respond.js doesn't work if you view the page via file:// -->
-    <!--[if lt IE 9]>
-      <script src="js/html5shiv.min.js"></script>
-      <script src="js/respond.min.js"></script>
-    <![endif]-->
-
-  <!-- jQuery (necessary for Bootstrap's JavaScript plugins) -->
-  <script src="js/jquery.min.js"></script> 
-  <!-- Include all compiled plugins (below), or include individual files as needed -->
-   
-  <script src="js/bootstrap.min.js"></script>
-  <script src="js/mp4.js"></script>
-  <script>
-	//load the menu
-	loadMenu();
-  </script>
-  </head>
-  <body>
-
-  <div class="container">
-	<!--the header root-->
-  	<div class="row">
-		<div class="jumbotron">
-			<h1>hi there, welcome to the mp4 site!</h1>
-			<p>the mp4 site is a simple webapp that can stream your own videos in responsive way.</p>
-		</div>
-	</div>
-	<div class="page-header">
-		<h1>all videos</h1>
-	</div>
-	
-	<!--row for link menu top-->
-    <div class="row">
-		<div class="panel panel-default">
-			<div class="panel-heading">
-				<h3>click any item to play, or click Play-All-[xx] to play all videos.</h3>
-			</div>
-			<div class="panel-body" id="medialinks1" ></div>
-		</div>		
-	</div>
-	
-	<!--row for the video display canvas-->
-	<div class="row">
-		<div class="panel panel-default">
-			<div class="panel-heading">
-				<div id="title">
-					<h3>select video to play</h3>
-				</div>
-			</div>
-			<div class="panel-body">
-				<div align="center" class="embed-responsive embed-responsive-16by9">
-				  <video id="video" controls class="embed-responsive-item">
-					<source src="" type="video/mp4" />
-				  </video>
-				</div>
-			</div>
-		</div>	
-    </div>
-
-	<!--row for link menu bottom-->
-    <div class="row">
-		<div class="panel panel-default">
-			<div class="panel-heading">
-				<h3>click any item to play, or click Play-All-[xx] to play all videos.</h3>
-			</div>
-			<div class="panel-body" id="medialinks2"></div>
-		</div>
-	</div>
-	<footer class="footer">
-		<p>© 2018 dmb</p>
-    </footer>
-  </div>
-  </body>
-</html>
